ulux-binding: Create interface UluxDatagram
