+++ bundles/binding/org.openhab.binding.ulux.test/src/test/java/org/openhab/binding/ulux/internal/EmptyDatagramMatcher.java
@@ -4,16 +4,16 @@ import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
 import org.hamcrest.Factory;
 import org.hamcrest.Matcher;
-import org.openhab.binding.ulux.internal.ump.UluxDatagram;
+import org.openhab.binding.ulux.internal.ump.UluxMessageDatagram;
 
-public class EmptyDatagramMatcher extends BaseMatcher<UluxDatagram> {
+public class EmptyDatagramMatcher extends BaseMatcher<UluxMessageDatagram> {
 
 	public EmptyDatagramMatcher() {
 	}
 
 	@Override
 	public boolean matches(Object item) {
-		return !((UluxDatagram) item).hasMessages();
+		return !((UluxMessageDatagram) item).hasMessages();
 	}
 
 	@Override
@@ -22,7 +22,7 @@ public class EmptyDatagramMatcher extends BaseMatcher<UluxDatagram> {
 	}
 
 	@Factory
-	public static Matcher<UluxDatagram> isEmptyDatagram() {
+	public static Matcher<UluxMessageDatagram> isEmptyDatagram() {
 		return new EmptyDatagramMatcher();
 	}
 }
