+++ bundles/binding/org.openhab.binding.ulux/src/main/java/org/openhab/binding/ulux/internal/handler/LedMessageHandler.java
@@ -13,7 +13,7 @@ import static org.openhab.binding.ulux.UluxBindingConfigType.LED;
 import java.util.Map.Entry;
 
 import org.openhab.binding.ulux.UluxBindingConfig;
-import org.openhab.binding.ulux.internal.ump.UluxDatagram;
+import org.openhab.binding.ulux.internal.ump.UluxMessageDatagram;
 import org.openhab.binding.ulux.internal.ump.messages.LedMessage;
 import org.openhab.binding.ulux.internal.ump.messages.LedMessage.Led;
 import org.openhab.core.library.types.DecimalType;
@@ -25,7 +25,7 @@ import org.openhab.core.library.types.DecimalType;
 final class LedMessageHandler extends AbstractMessageHandler<LedMessage> {
 
 	@Override
-	public void handleMessage(LedMessage message, UluxDatagram response) {
+	public void handleMessage(LedMessage message, UluxMessageDatagram response) {
 		for (Entry<String, UluxBindingConfig> entry : getBindingConfigs(LED).entrySet()) {
 			final String itemName = entry.getKey();
 			final UluxBindingConfig config = entry.getValue();
