+++ bundles/binding/org.openhab.binding.ulux/src/main/java/org/openhab/binding/ulux/internal/handler/PageIndexMessageHandler.java
@@ -14,8 +14,7 @@ import static org.openhab.binding.ulux.internal.UluxBinding.LOG;
 import java.util.Map.Entry;
 
 import org.openhab.binding.ulux.UluxBindingConfig;
-import org.openhab.binding.ulux.UluxBindingConfigType;
-import org.openhab.binding.ulux.internal.ump.UluxDatagram;
+import org.openhab.binding.ulux.internal.ump.UluxMessageDatagram;
 import org.openhab.binding.ulux.internal.ump.messages.PageIndexMessage;
 import org.openhab.core.library.types.DecimalType;
 
@@ -26,7 +25,7 @@ import org.openhab.core.library.types.DecimalType;
 final class PageIndexMessageHandler extends AbstractMessageHandler<PageIndexMessage> {
 
 	@Override
-	public void handleMessage(PageIndexMessage message, UluxDatagram response) {
+	public void handleMessage(PageIndexMessage message, UluxMessageDatagram response) {
 		LOG.debug("Page {}", message.getPageIndex());
 
 		for (Entry<String, UluxBindingConfig> entry : getBindingConfigs(PAGE_INDEX).entrySet()) {
