+++ bundles/binding/org.openhab.binding.ulux/src/main/java/org/openhab/binding/ulux/internal/handler/VideoStateMessageHandler.java
@@ -13,7 +13,7 @@ import static org.openhab.binding.ulux.UluxBindingConfigType.VIDEO;
 import java.util.Map.Entry;
 
 import org.openhab.binding.ulux.UluxBindingConfig;
-import org.openhab.binding.ulux.internal.ump.UluxDatagram;
+import org.openhab.binding.ulux.internal.ump.UluxMessageDatagram;
 import org.openhab.binding.ulux.internal.ump.messages.VideoStateMessage;
 import org.openhab.core.library.types.OnOffType;
 
@@ -24,7 +24,7 @@ import org.openhab.core.library.types.OnOffType;
 final class VideoStateMessageHandler extends AbstractMessageHandler<VideoStateMessage> {
 
 	@Override
-	public void handleMessage(VideoStateMessage message, UluxDatagram response) {
+	public void handleMessage(VideoStateMessage message, UluxMessageDatagram response) {
 		// video active
 		for (Entry<String, UluxBindingConfig> entry : getBindingConfigs(VIDEO).entrySet()) {
 			final OnOffType newState = message.isVideoActive() ? OnOffType.ON : OnOffType.OFF;
